changed class name and fixed bugs
