+++ nf_core/pipeline-template/{{cookiecutter.name_noslash}}/main.nf
@@ -13,25 +13,18 @@ def helpMessage() {
     // TODO nf-core: Add to this help message with new command line parameters
     log.info nfcoreHeader()
     log.info"""
-
     Usage:
-
     The typical command for running the pipeline is as follows:
-
     nextflow run {{ cookiecutter.name }} --input '*_R{1,2}.fastq.gz' -profile docker
-
     Mandatory arguments:
       --input [file]                  Path to input data (must be surrounded with quotes)
       -profile [str]                  Configuration profile to use. Can use multiple (comma separated)
                                       Available: conda, docker, singularity, test, awsbatch, <institute> and more
-
     Options:
       --genome [str]                  Name of iGenomes reference
       --single_end [bool]             Specifies that the input is single-end reads
-
     References                        If not specified in the configuration file or you wish to overwrite any of the references
       --fasta [file]                  Path to fasta reference
-
     Other options:
       --outdir [file]                 The output directory where the results will be saved
       --publish_dir_mode [str]        Mode for publishing results in the output directory. Available: symlink, rellink, link, copy, copyNoFollow, move (Default: copy)
@@ -39,7 +32,6 @@ def helpMessage() {
       --email_on_fail [email]         Same as --email, except only send mail if the workflow is not successful
       --max_multiqc_email_size [str]  Threshold size for MultiQC report to be attached in notification email. If file generated by pipeline exceeds the threshold, it will not be attached (Default: 25MB)
       -name [str]                     Name for the pipeline run. If not specified, Nextflow will automatically generate a random mnemonic
-
     AWSBatch options:
       --awsqueue [str]                The AWSBatch JobQueue that needs to be set when running on AWSBatch
       --awsregion [str]               The AWS Region for your AWS Batch job to run on
@@ -60,7 +52,7 @@ if (params.help) {
 def json_schema = "$baseDir/nextflow_schema.json"
 def unexpectedParams = []
 if (params.validate_params) {
-    unexpectedParams = Schema.validateParameters(params, json_schema, log)
+    unexpectedParams = NfcoreSchema.validateParameters(params, json_schema, log)
 }
 ////////////////////////////////////////////////////
